overhaul scrapers/math_dept.py (#91)
This is major overhaul of `scrapers/math_dept.py`. Even though I know it
isn't actually being used in production, I still think it's good to make
all the code as pretty and modular as possible for posterity.

More specifically, the main code in `math_dept.py` (all the random stuff
that used to be non-indented) is now divided into:

* `run()`, the main entry point, which uses three other helper
functions:
  * `get_rows()` for scraping
  * `parse_subject(subject)` to parse the title specifically
  * `parse_row(row)` to parse an individual row
* `test_parse_when()` as a sanity check that `parse_when(when)` works
properly