move off of cra to vite (#96)
Working with CRA is a bit of a pain and there's been work to move off of
it. I didn't update #57 since that involves the creation of a backend
server too (which involves setting up XVM and all that) and this is more
of an instant change, although rebasing that PR too would be good (maybe
over IAP when ppl have more time?).

This PR moves off of CRA and moves to Vite. This also involves having to
recreate some of the tooling integrated into CRA, such as an eslint
config (a lot of the diffs in component files were issues flagged by
tseslint that weren't being checked before). I didn't really have to
change the workings of any components, which was nice. This issue also
resolves all the security vulnerabilities found when running `npm
install`, and also closes #89.

Doing this would probably make anything we want to do in the future,
such as adding tests (which can be done with
[Vitest](https://vitest.dev/)) or integrating a backend server (which
can be done using [a
plugin](https://github.com/axe-me/vite-plugin-node)), a bit easier.