Include classes that lack schedules + split off IAP/summer (#100)
Fixes #27 and #7. In the scraper, replace the check that a class has a
schedule with a check that a class is offered in the current term (with
the fall/IAP/spring check done in `catalog.py` while the academic year
check is done in `fireroad.py`), and add proper support for IAP and
summer terms.

---------

Co-authored-by: Diego Temkin <65834932+dtemkin1@users.noreply.github.com>