Update course 18 underground guide link to new site (#108)
Fix broken link to the course 18 Underground guide. The underground
guide is at [mathguide.mit.edu](https://mathguide.mit.edu) now. To view
a specific class, go to, e.g.,
[mathguide.mit.edu/18.404](https://mathguide.mit.edu/18.404).

It used to be at [course18.guide](course18.guide), but the domain is no
longer active. As a maintainer of the underground guide at [MIT UMA
Tech](https://uma.mit.edu/resources), we work with
[CoMM](https://web.mit.edu/comm/) to rewrite the site. Although the work
is still in progress, it'd be great to update the link to ease class
selection!