Create form for course admins to submit overrides (#185)
Fixes #70. The UI is a little janky and isn't consistent with the main
Hydrant site, but (while we can/should smooth out rough edges) I think
using [JSON Forms](https://jsonforms.io/) for this is appropriate since
it avoids duplicating the contents of `rawClass.ts` in yet another
place, and this page should only be seen by course administrators anyway
(and not by general Hydrant users).


![image](https://github.com/user-attachments/assets/ee5fac25-5bc5-4902-93dc-d122341521ef)

The "submit" button opens a mailto link and populates the body with TOML
data (mailto doesn't support attachments).


![image](https://github.com/user-attachments/assets/657a43ff-a16e-4f4e-a394-e771eccae750)

A known issue is that the initial state includes a "must be array" error
since the data object is initialized to `undefined` instead of the empty
array; I haven't been able to figure out how to fix this.

![image](https://github.com/user-attachments/assets/76539ed5-63cc-483a-be98-95e07d80b3ae)

---------

Co-authored-by: Diego Temkin <65834932+dtemkin1@users.noreply.github.com>