Separate CI pipelines for frontend and backend (#209)
This PR eliminates the old `.github/workflows/ci.yml` file in favor of
two new pipelines, which are defined in
`.github/workflows/ci-frontend.yml` and
`.github/workflows/ci-backend.yml`. The former handles Prettier, ESLint,
and typechecking. The latter handles Black and Pylint. It also
implements a rather naïve integration test wherein `public/latest.json`
is checked against a JSON schema (which I put in
`scrapers/hydrant_schema.json`, for lack of a better folder) that I
specifically cooked up just for this PR. (We don't unit-test the
backend, but I also think this could be overkill since the code is
fairly simple and `public/latest.json` is the only output that actually
matters.) Furthermore, I designed the frontend pipeline to *not* trigger
on backend-only changes, and vice versa, and *neither* should trigger on
documentation-only changes.

If accepted, this new pipeline (if it works; we'll have to see) could
lead to clearer feedback, since many pull requests and commits are
frontend-only or backend-only, making the other checks spurious. The
integration test, furthermore, should prevent most functionality-related
bugs on the backend from making it into production. Frontend testing
will have to wait, though; I'm thinking of making this my next PR.